pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "noid-rag"
version = "0.1.1"
description = "Production RAG CLI & Agent Skill"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "docling>=2.0.0",
    "httpx>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "tenacity>=8.2.0",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
local = ["sentence-transformers>=3.0.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "pytest-cov>=5.0.0", "ruff>=0.4.0"]

[project.scripts]
noid-rag = "noid_rag.cli.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/noid_rag"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
